+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# 1.1.0
+- General
+  - Moved all login credentials to `Edit->Preferences->Steam Build Uploader`
+  - Added icon to the window
+  - Logging is a little better
+- Steam tab
+  - Changed lists of depots and branches to be reorderable
+  - Changed default branch to be called "none" instead of "default"
+  - Added button to open steam sdk console
+  - Added button in configs to open the games steam store page
+  - Added button to open steam sdk page to download it if your path is not valid
+- Unity Cloud tab
+  - Fix for macos builds spamming errors
+  - Builds older than 1 year now show the exact years
+- Upload tab
+  - Default source type changed to Manual
+  - Manually uploads now require a `.exe` or `.zip` in order to upload. `.zip` will be unzipped when uploading.
+  - Builds will no longer upload if any sources failed to provide a valid path 
+
 # 1.0.1
 - Replaced EditorCoroutine with Tasks (You can remove that dependency package now)
 - Hid StreamSDK and UnityCloud credentials